Write to ApplicationWorkHistoryBreak#breakable
ApplicationWorkHistoryBreak is going to be a polymorphic table, we want
it to belong to an application_form or an application_choice.

This will help us make the application work experiences editable on the
application form.

This commit will save the link between the
application_work_history_break and the application_form in the
polymorphic columns, another commit will follow to back-fill the data