Ignore ApplicationWorkHistoryBreak#application_form_id
We don't need the application_form_id on the
application_work_history_break table, we use the breakable column to
link it to an application form.

This commit ignores the application_form_id and makes sure the deletion
of the history break is going to happen when deleting a candidate, just
like before.

We made application_form_id nullable here
9d60480d2c5c74222a3168de6908431ab185f438