Use application_choice polymorphic associations
In the future, every application choice will have an
application_experience or application_work_history_break. Through the
exprienceable or breakable polymorphic columns.

This will allow the user to edit work experienceas or work history
breaks for new application choices.

To enable this feature we must get the work expeirences and histories
through the application_choice on the provider interface and vendor api.

We would look if the application_choice has the data we need and if not
we go to the application_form.

Eventually when we migrate the data making every application_choice have
work experiences or work history breaks, this logic will not be needed.