Dup work experiences and histories on choice
To allow the user to edit the work experiences or histories we will
duplicate  the experiences and histories on the application choice from
the application form.

This will allow the user to edit the experiences or histories on the
application_form and the edits will be applied to future application
choices.

The current application_choices will have the old info

We also don't want to update the application_choice if the
application_form work experiences or histories have been updated