Fix /provider/activity view
With the new structure of the application experiences and
application work history breaks.
There are more types of audits being logged. This
broke the `provider/activity` view where the provider
sees what the user has done before submitting the application.

This commit will skip these logs for the moment.

This is a live issue so we're tying to buy some time until we make a
decision if we need to display this information or not.