Exclude work experiences from provider audits
With c5c12877533088397b8197de742204bde6028a2a we save
application_experiences and application_work_history_breaks for the
application_choice. Part of allowing the user to edit work experiences
and histories.

This creates audits that the `/provider/activity` view tries to show.
These audits don't need to be shown to the provider. The work histories
and breaks cannot be edited on the application_choice, they can be
edited on the application_form.

So the provider doesn't need to see any audits related to work histories
and breaks on the application_choice.

This commit removes these audits from the sql query for this provider
view