Fix experiences & histories when submitting choice
There seems to be a very rare edge case when submitting an application
choice. If the choice has work experiences or histories already we cannot
dup the work experiences and histories from the application_form.

This commit fixes this by removing any existing work experiences and
histories from the application choice one by one before duping the
application form ones.

Because we set the work experiences on the choice with the `=` sign.
This will try to delete all records existing and set new records. But
there's a `PG::NotNullViolation` on the `experienceable_type` that stops
the `delete_all` method. We need to delete a record one by one.