Add new field to application forms
We need to enable candidates without a degree to apply for TDA courses.

We want them to be able to mark all section as complete in the task list,
so they can feel confident they have done everything they need in order
to apply for courses.

This field will represent the answer from the user on the new first
question on the degrees for the next cycle.
"Do you have a university degree?"
Yes
No

If no the section will be marked as complete.
If yes they will go to the current degree wizard.