Add new question of university degree to degrees in the new cycle
We need to enable candidates without a degree to apply for TDA courses.

We want them to be able to mark all section as complete in the task list,
so they can feel confident they have done everything they need in
order to apply for courses.

The feature:

The page view is updated to include a question if they have a degree or not
If they have a degree the flow proceeds to what is currently live
If they do not have a degree, they are taken back to the tasklist
with the section marked as complete