Allow candidate to edit work histories
We want to allow the candidate to edit their work histories, this will
only edit work experiences, volunteering_experiences and work history
breaks on the application form.

The application choices that are submitted will still have the old work
histories. This is because each submitted application choice has its own
work history.

So the user will be able to edit their work histories but only for new
application choices.