Add flight check for undergraduate courses when candidate has a degree
TDA is an undergraduate route into teaching, and applications are
primarily for those who do not have a degree.

Those who have a degree are still able to apply, however there are
other quicker and more suitable routes into teaching for them.

This commit is to add a flight check to TDA applications where
the candidate has added at least one degree, to encourage them
to consider alternatives.