Improved delete query
The BatchDelivery is asking for all the columns of the audit table
in the sql query. We are interested in only the id.

Also it does a count on all the records which we don't want as we have
about 3 million records that need to be removed and counting can be
slow.