Change work_histories migration to insert via SQL
Previously, the data migration
BackfillApplicationChoicesWithWorkExperiences was inserting records via
ActiveRecord. This caused an incident, known as 'The Big Touch'. A lot
of application choices, forms and candidates had their updated_at
updated and this caused some of our APIs to be slow.

We still need to run this migration in sandbox so we thought it's a good
idea to change this migration to insert records via SQL avoiding any
callbacks.