Fix casting bug when backfilling work histories
There are two enums on the ApplicationWorkHistory record. part_time and
full_time. ActiveRecord returns these enums in the console like so
'part_time' but in the DB it's saved `Part time`.

Previously because we called `record.attributes` it returned 'part_time'
and the value in the DB would be saved as 'part_time'. Not 'Part time'.

This would break some views as ActiveRecord expects the value in the DB
to be 'Part time'. This commit fixes this issue.