Change updated at of application choices
This PR attempts to fix some of the issues caused by The Big Touch

It will revert all the application choices that have been touched to the
previous updated_at they had. We do this using their audits.

We get all the choices on the 3rd of September 2024 and get the touch
audits for each of these choices and if there is an audit before the
touch moment we revert the choice to the created_at of that audit.

If the choice has user created audits after the touch, we don't update
that choice.

The data migration allow us to pass a limit or an array of provider ids
to target a provider choices or to run the migration on a limited amount
of choices. These params are optional.