Add link to FutureLearn style acceptance tests
The testing style guide used to link to a blog post on the FutureLearn
website. It detailed an approach to writing Cucumber-style acceptance
tests using RSpec.

The blog post has since been removed from the FutureLearn website.
However it's still available thanks to the Wayback Machine on
archive.org, so it can still be linked to and referenced in this
document.

The blog post itself still seems to be useful and relevant, so I think
it's worth re-adding the link but using the archive.org version.