Fix incorrect candidate bachelor degree equivalency
When a candidate inputs their degrees, they can chose an option called
'Another qualification equivalent to a degree'. This free text field
will try to figure out if the value imputed is a bachelor degree or not.

Previously we would just check if the free text search value is in our
list of Bachelor degrees by checking if the string is included in the
list. We didn't match the strings exactly.

So if you have the correct keywords, like Bachelor of Medicine and Pixie
hunting we would classify this degree as a bachelor degree because there
is a valid bachelor degree called `Bachelor of Medicine`.

This caused the form to break as there is no valid degree for Pixie
hunting in the reference data gem.

This commit checks the strings exactly to avoid this issue.