Fix hesa ethnicity for 2023/2024 application forms
In this commit 05c8752f1f5beee512278f735029036b77eb9479 we have fixed a
bug where the `Another White background` option to
`ethnic_background` saved the wrong `hesa_ethnicity` code. It saved 160
rather than 179.

We now need to fix this issue for the old application_forms. There are
about 5k application_forms in production, only for 2023 and 2024
recruitment cycles. This commit adds a data migration to fix this.