Bring the journey uk and non uk degree in line
Currently when adding a UK degree, the first step after choosing the
country is the type of degree.

When adding a non UK degree, the first step is the subject of the
degree.

This commit changes the non UK degree journey to so that the first step
is the type of degree. Bringing both journeys steps in line