Fix degree form when chaining countries and backing off
There's an edge case when editing a candidate edits a degree

If they edit the country and click the back link and select the old
country they already have and save. The form would have cleared the
degree type, subject and university information.

Asking for the user to input this information again is valid as the
degree types, etc don't match when you change the countries.

But if you go back through the form and select the country you
originally have, the form should not erase your degree type, subject and
university choices.

This commit fixes this. By validating that the data is there before
saving it to the ApplicationQualification record.