Add choice deferred timestamps for tad report
We want to add 3 new fields to the tad report.

offer_originally_deferred_at - The timestamp of the first deferral offer
offer_reconfirmed_at and - Timestamp when the deferred offer has been
confirmed by the provider
offer_reconfirmed_cycle_year - The recruitment cycle the confirmation of
the deferral falls in

The information for these fields needs to come from the audits table,
there's no other way to get this information at the moment.