Updated wording for "QTS with PGDE" qualifications (#9987)
## Context

The text "PGDE with QTS" does not match up with other services and
should be reworded to "QTS with PGDE".
This is inline with the Find & Publish services and the GiT website. 

## Changes proposed in this pull request

We are using `gsub` in the description to ensure previous Courses
receive the updated wording without migrating. This matches the
implementation of the "QTS with PGCE" rewording.

Both `description_to_s` and `qualifications_to_s` methods are used only
for displaying text.

## Guidance to review

- Check usages