Add safe choice touch validation in support
Often when a support user tries to update an application form that is in
the previous cycle, we raise an exception. This exception was not
handled so the page would brake

This is because when updating some specific columns on the application
form, we touch the choices of that form.

If the application_form is in the previous cycle, it's not deferred or
the global variable allow_unsafe_application_choice_touches is not true
we don't allow the support user to update the application form, by
showing an error to the support user

This is because the provider would not get these updates, so we don't
want them to happen.