Add rescue step to degree form
The degree form can error if the redis data store is cleared when users
are using it.

This can happen when a user uses multiple tabs to input their degrees.
After saving the changes in one tab, we clear the redis store so the
other tabs would get an error when saving the degree form.

To fix this a much bigger refactor is necessary.

This commit tries to gracefully handle this edge case, so that the user
doesn't see an error.

It will redirect to the review page if the user has a degree or to the
beginning of the degree form when inputting a degree for the first time,
in case the form fails when adding the first degree.

We still raise an error in sentry so we don't lose sight of this issue.