Delete big touch audits
When we ran the big touch, on the 3rd of September 2024, we created
about 3 million audits that are not necessary.

This commit deletes these audis.

We only intend to delete audits for ApplicationExperience and
ApplicationWorkHistoryBreak that are associated with ApplicationChoice,
these audits don't have a user attached to them and the username
'(Automated process)'