Remove default footer
We were using a footer to render the Get help section, but in practice, it is not always the last thing on the page and often emails have bespoke get help text. We were adding endlessly to the exceptions. This commit composes the emails with the get help section where it belongs rather than relying on a footer.

clean up