Change the closed course message
Removed the 'now' word from our closed course message.

This should be a temporary change. Providers want to show their TDA
courses before they are published in find, they can do this by having
the course closed but this means that the course will be available to
apply in apply, though the candidate will get the closed course message.

We want to remove the word 'now' from the closed course message because it
indicates the course was available in the past but it's no longer when in fact,
for TDA courses is more like they are not available yet.