Fix back link for references interruptions
Sometimes the reference is not present in our DB any more. This can
happen if the user deleted the reference in another tab.

The back link for the references#interruptions controller broke when the
references didn't exist any more, now it will just redirect the user to
the references index page.