Update smoke test to work with one login
The smoke test needs to be smarter now that we will introduce one login.

We need to check if one login is enabled but we don't have control over
the host app, as we use a remote host, when we deploy to review we use
the review app host, same with deploying to other envs.

This commit tries to make the smoke test smarted and tests one login or
magic link depending on what is the sign up method on the host.