Add DB backed session tables
In testing the one login integration, we discovered that having session backed
login is not ideal. There is a risk that the session gets too big, over 5000 bytes.
To fix this, we are switching to Database backed sessions.

We will use DB backed session for the normal candidate flow.
Impersonation will still use cookie session.

This means that deleting the session from the DB logs out the user.

I've also added a SessionErrors table, this will hold all the errors that
are generated by our one login integration. Storing errors in DB makes it
easier to debug, as we store the omniauth hash and all the details for a
particular user. This removes the need to push sensitive information to Sentry.