One login account recovery feature
When going live with one login, a candidate can sign up with a different
email address, not the magic link email address.

For this, we need to allow our candidates to recover their 'old'
account. This commit adds this feature.

We show a banner which the candidate can dismiss or they can click to
recover their old account. They will be asked to input their old email
and a code will be sent to their email. The code is encrypted with
bcrypt.