Rename Account Recovery Request Code
We need to use has_secure_password for the code on the Account Recovery
Request Code. To use this helper, you need to have a xxx_digest column
in DB.

This commit creates this column. There are no records in any environment
for this table, using safety assured should be enough.