New header and primary navigation for candidate support and provider
With the introduction of one login and the fact that we need a new
header and primary navigation for the candidate path, to allow the
candidate to manage their one login account.

We decided to implement the header and primary navigation for our other users.
providers and support as well as candidates.

I have not changed these components for our api users. The content
that's already there needs a bit more work before, I think.

This should show the GOV.UK One Login account link for candidates if the
one login feature is enabled.