Hide account recovery banner to users without one login auth
There is an edge case when one login will be switched on. The candidates
that are already logged in, will still be logged in.

The issue comes when the account recovery banner will appear and when
the one login feature flag is enabled, the user won't have a one login
auth record in the DB. This record links the one login account to a
candidate account.

If this one login auth record is not present, you cannot recover an
account as the recovery account feature will look for a one login auth.

The 'fix' for this is to allow instruct the user to log out and login
with one login, this will create a one login auth and then they can
recover their account without issue.

This commit doesn't allow the candidate to recover their account unil
they have a one login auth record in DB.