Fix candidate forms using current_user
Some forms like the degree form use current_user. In the candidate
context, this is the same as current_candidate.

As we switch to one login and DB backed sessions, current_user will be
nil. So to make this work and make it backwards compatible, in case we
need to revert back to using magic link for our candidates.

I've overridden the `current_user` method in the candidate interface
controller to always point to current_candidate. This way it should work
for DB backed sessions and cookies backed sessions.

I've also changed the degree system specs to use one login when signin
up. We would've probably spotted this if the spec was using one login to
sign up the candidat