Switch candidate system specs to use one login
We are close to releasing one login and it would be good if we switch
our way of signing in the candidate in candidate system specs.

With one login, we also introduced DB backed sessions. As this becomes
the normal way to sign in. Our system specs should replicate this
behaviour. This way we can catch bugs easier.