Add migration to SessionError table
There are two types of errors the candidate can encounter when using one
login. Typical 500 errors and when a user uses the wrong email address

We need to know what error the candidate gets to redirect them to the
correct page.

Prior to this change we are redirecting the candidate to wrong_email
error page even if it's a 500 error page.

Adding a error_type to the SessionError table makes it easier to
distinguish between these errors. Better than storing this information
in the session.