Redirect the candidate to the correct error page
There are two types of errors the candidate can encounter when using one
login. Typical 500 errors and when a user uses the wrong email address

We need to know what error the candidate gets to redirect them to the
correct page.

Prior to this change we are redirecting the candidate to wrong_email
error page even if it's a 500 error page.

This commit redirects the user to the correct error page