Fix sandbox page redirect
On sandbox, when we log out the user. We redirect them to
contents#sandbox.

This controller is not properly setup to deal with one login, especially
when the one login feature is turned off.

Redirects from the page of this controller hit the Authentication. The
concern returns true that the candidate is signed but with db backed
session, not devise. So the user is stuck in a loop because devise
thinks the user is not signed in but our DB backed session concern says
it is.

Clearing the session fixes this issue.

In reality this should not happen because we don't attempt to use the db
session login when one login is off, in any of our other controllers.
But requests from this controller accesses the Authentication concern.

This commit tries to fix this by just clearing the session if one login
feature is not enabled, this will fix this issue. Ideally we would want
to not have this controller send requests to the Authentication concern
if one login is not enabled.