Log current_candidate as audit user in candidate_interface
With the introduction of one login we accidentally introduced a bug in
our audits.

Some of our audits have the user_id nil because the alias that sets the
audit user was on the wrong line in the candidate_interface controller.

This commit fixes this. I've also added a spec to make sure this doesn't
happen.