Candidate pool tables
To have a working candidate pool for providers we need to allow
providers to invite candidates to apply to courses.

We also need candidates to opt in. And a temporary list of providers
that opt in as well.

This commit adds all the migrations that we need for this feature