Implement params.expect along with rubocop rule
params.expect is a much safer and cleaner way to define what params a
controller should receive. This is a feature of rails 8. The old way is
still supported but we want to upgrade the way we handle params in our
controllers.

Some controllers like the vendor api will be left as they are because
`params.expect` will change the response when the params are wrong