Fix CI enum check
The pull request checks that we run in ci have a bug.

When cheking if an enum has changed in the code. We pull the branch, run
db:setup and save the enums to a file

We then checkout main and at this point we have a problem. Because we
checkout main, we lost the changes in analytics yaml, so dfe analytics
can raise an error at this point because the DB was setup in the feature
branch not main.

This commit tries to fix this by setting the DB again after checking main
so that the columns match what's in analytics yaml on main.