[653] Reset grade and constituent_grades when qualification type changes
There was a bug where UK grades appeared in the grade section on the review page when a qualification was changed to non-UK. Clicking the "Change" link did not surface any values.

It makes sense to clear the grades when changing qualification types so the user has to explicitly enter them.