Candidate pool filters geolocation filter
This commit adds 2 filters to the candidate pool index page.

Location based filter and ordering and visa sponsorship.

I've used the geokit rails gem to calculate the distance from the
filtered location to the actual site location.

The location filter also makes calls to Google maps autocomplete api to
help the user find postcodes cities, etc. Sadly this endpoint doesn't
return lat and long of each location.

So another api call will be done when you apply the filters to the
google maps places details endpoint to get the lat and long which are
used to get all the sites within a range of a location.

Once you select a location filter, we order the results based on the
proximity of the sites to that location.

These api calls are cached.

I've also changed the Pool::Candidate to return application forms. This
way is easier to query and read the queries, I think and also all the
information for the index page is on the application form.