Persist find a candidate filters in DB
We want the filters for find a candidate to persist from page to page.

To do this we need to store them somewhere. We don't really want to
store them in the session because our sessions can overflow. We already
store some data in the session and storing more is a risk.

This PR stores the filters on the provider user table. in a jsonb
column. Hopefully we will not need to query this data.