Updated Notify Check to use the GENERIC_NOTIFY_TEMPLATE ID (#10538)
## Context

We're making this change as the Test Template was deleted, causing our
status to report as down, even though the service is fully operational.
We have chosen to use the generic template as we would like to be
alerted if this template is ever deleted, which would better reflect the
service being down.

## Changes proposed in this pull request

- Change test template id is health check

## Guidance to review

- N/A