Candidate preferences form v1
This is a version of the candidate preferences form.

It's not the final version, changes will come after this, allowing the
candidate to not have any location preferences if they want and allowing
the candidate to come back to their preferences and update them.

We also need a way to integrate this form into the candidate flow, right
now the only way to get to this form is to know the url.

This is just a basic/complex CRUD, multi step form.

We have a controller and form object for every step. This makes it
easier to handle the redirect, back buttons and save/update logic.

We do pass around a `return_to` param to indicate to the form where to
return. Used on the review page.

This tries to follow REST rails controllers.