Share details page
We need a page to funnel users to add their candidate preferences.

This page is the share_details page. It will be presented to the user
after submitting an application choice. Only if the
`candidate_preferences` feature flag is active and if they didn't setup
their preferences before.

From this page, they will be redirected to the candidate preferences
form.

In this commit we also add a `manage_preferences_component` to the
bottom of the applications index page. Where the user can edit their
preferences or view the guidance of the share details page again.