Edit candidate published preferences
This allows the candidate to edit their published preferences.

This is different from the flow of creating/editing draft preferences.
Because we are editing published preferences.

There's a new publish_preferences show controller action, which is very
similar to the draft_preferences show action. They are separate by
choice to keep the logic separate.

The change buttons go to the same controller actions, pool_opt_ins or
draft_preferences controllers but we dup the published record to a draft
record and continue editing the draft record until we publish the draft.

Once we publish the draft, the old published preference is deleted