Improve location preferences form and location suggestions
This validates the location imputed by the user in location preferences.

Determining what's a real location and what's not is hard.

Geocoder.search can return places that are not actually real locations.
But the LocationSuggestions service uses the autocomplete google api.

This returns real places. So we can use this endpoint to validate the
user input.

We take the first suggestion the user imputed and save it in the
database. This means that if the user Added 'M20' the value in the DB
will be Manchester M20.

If the user imputed 123 there won't be any suggestions so we'll show an
error.

This way we can guarentee that whatever location is ending up in the
database is a real location and has lat/long values.

We use the place_id from the location to get the coordinates