Add provider to location preferences
When the user opts in to the find a candidate feature. We will generate
some location_preferences for them. These are based on application
choices and home address.

The ones generated from application choices need to show the provider
name next to their name. Previously we put this in the name of the
location preference. But this is not a good idea.

So a location preference can belong to a provider, this is optional.

When we generate these location_preferences we link them to a provider
and use that provider to display the provider name next to them.

A user can also update a generated location, if they do, we need to
remove this link, the link to a provider so that the provider name is not
displayed next to the location anymore.