Stop opted out candidates from viewing the preferences review page
We don't really want candidates who opted out of the find a candidate
feature to get to the review page of the candidate_preferences.

This commit adds some logic to redirect the user to the edit opt in page
if they have a published preference with an opt out option.

Effectively prompting them to opt in to get to the review page.