Add candidate id to find a candidate feature
Providers have raised the point that is hard to keep track of what
candidates they invited if the cannot see the name of the candidate.

So exposing the candidate id will help them with this.

Providers using the api already see the candidate id. So they will only
be able to identify their own candidates in the find a candidate list

This is not an issue