Dynamic location preferences
When candidates Candidates opt in into the find a candidate feature and
set their location preferences to be updated as they apply for other
courses(:dynamic_locations).

We need to update their location_preferences with these new locations.

This commit adds a LocationPreferences service that can add the default
locations when creating the preferences but also add 1 location to the
candidate preference.

This is called in the submit application choice service.

To not add duplicate locations we check the names off the location
preferences against the new site postcode.