Preserve the order the location preferences are created in
Because of the way our multi step form works for the candidate
preferences. The order of the location preferences would change.

This is because we dup the published preference into a draft preference
when the candidate edits a published preference.

The dup method was using insert_all which is performant but will result
in all the location preferences having the same created at. So the order
would be lost.

This changes that and also stops updating the location preference name
with the suggestion name if the name of the location preference has not
changed.