Change the Pool::Candidate query to use candidate preferences
Previously the query that was used on the find a candidate index page
was looking within the sites of the candidate application choices.

This needs to change. We need to use the candidate preferences to show
results on the find a candidate view, in the provider interface.

This changes the query. It calculates the distance from the provider
location to each candidate_location_preference and returns the first result if
any of the location_preferences are within distance.