Change content for candidate preferences
The CandidateInterface::ShareDetailsComponent needs to behave a bit
different depending how the user gets to it.

If they see this component after submitting an application the content
is different than when they come back to this component to read about
how the find a candidate feature works.

We do this through a param in the url