Push candidates that are in the pool to BigQuery
We want to have a historical view of what candidates where in the pool
and how they were added and removed from it.

We know this is dynamic and can change quite quickly but even a snapshot
on a daily basis can be helpful in building a picture of this.

The PoolEligibleApplicationForm tables is not really meant for us to
use. This table is meant for our data team.

For now we'll run a daily job. And they can use the application_form_id
and created at to build their dashboards.