One login redirect to request referer
Previously, when you would try to access a candidate url and you were
not logged in. You would be prompted to login but we would not
redirect you to the path you were trying to go.

We would just land you on your details or applications page.

This change will redirect the user to wherever they wanted to go.

We put the path we want to go in the url before sending a request to one
login. This path comes back to us in the callback method. This path is
set in our Authentication concern. Just before we ask the user to login
if they are not login.

We can use this path to redirect the user to wherever they wanted to go
after we confirm their one login.

Before we do this though, we need to check that the path value is an
actual path in our app. We use Rails.application.routes.recognize_path
for this. This will return the errors controller with the not_found
action if the url/string value is not a valid route within our app.

If the url is valid we redirect the user to that url. Otherwise we
default to the previous behaviour of redirecting the user to 'Your details'