Added worker to send FAC feature launch emails (#10598)
## Context

We want to have a worker that can take a list of Application Form ids
and send the FAC feature launch email to a sample of 1000 Candidates.

The emails will initially be sent to a sample of 1000 Application Forms
- the list of application_form_ids will be supplied by the Data team.
Checks for eligibility, previously sent and locked, blocked and
unsubscribed will take place before limiting the sample to 1000.

Candidates will be sent a variant of the
`find_a_candidate_feature_launch_email` which they will automatically be
assigned to.

For tier 3 Candidates, we will remove the
`.merge(Pool::Candidates.application_forms_eligible_for_pool)` as we
want to target all Candidates which have submitted an Application

## Changes proposed in this pull request

- Added FindACandidateFeatureLaunchEmailWorker