Fix location preference show view
There's an edge case where a user can delete a location preference in
one browser tab but on another browser tab, this location preference
might still appear because the user didn't refresh the page.

This commit redirects the user to the location preferences index page if
there's not a location preference found in DB. Fixing this issue.