Redirect to application choices if the candidate preference is blank
There are edge cases, probably from creating candidate preferences in
separate tabs where we the ids of the preferences can get lost because
records are being updated/removed in a different browser tab.

This adds some guard rails, if the form loses the preference record it
redirect the user back to the application choices page.

This fixes the rare edge cases where the in the url doesn't what we have
in DB anymore.