Handle candidates with no location preferences in the pool
The candidates that have no location preferences but they did opt in
should still be in the pool. They should be bumped up the list and
instead of showing how far are from a location, we show that they are
open to any course in the country.

To do this, we change the filter_by_location query to do a left join,
which returns the parent table we are querying. In this case
candidate_preferences. If the candidate has a published preference and
they opted in. They will appear in the filter_by_location query, with a
distance of 0.0. This puts them at the top of the list.

Where we can change the label based on some logic to say that these
candidates are open to relocation.