Scope the candidate_pool query to current cycle and return less columns
It turns out that in some of our subqueries we were not scoping the
application_form to the current cycle.

Also, when we return all the columns of the application form, ordering
them is slower then returning only the columns we need for the view.

These 2 changes have almost half the query time. On my local.

With 82k candidate_preferences.
Before the change the query would take about 1.6-1.7 seconds after it takes
around .7-.8 seconds

We can further improve this with an index on the recruitment_cycle_year
on the application form. That will follow in a different commit