Add recruitment_cycle_year index to application form
In my testing, with 82k candidate preferences an index on the
recruitment_cycle_year would improve the query by about 100-150ms

Not much but it can add up. I did this on my local so it's not the most
accurate thing but I don't see a harm in adding this index.