FAC location search query fix
In trying to add people with no location preference to the candidate
pool. We've accidentally included everyone that opted in regardless of
the distance, when searching by location.

This is because of the left join we did, which returns all the candidate
preferences that opted in.

This commit fixes this by removing the left join and adding 2 sub
queries in union to the join with candidate preferences

In the first subquery we get all the candidate location preferences
within the range of the searched location and in the second we get all
the opt in candidate_preferences without location preferences.

This way we include both candidates within range but also candidates
without a location preferences.

We set -1 to the candidates without location preferences to make sure
they come at the top. 0.0 was not good enough as there was a possibility
that another candidate with a location preferences would be within 0.0
of a location.